Move networking to new network module.
