+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![allow(dead_code)]
 
 pub mod colour;
+pub mod network;
 pub mod request;
 pub mod response;
